Restructured filetreepanel and filetreepipeline.
