+++ web/editor.html
@@ -113,14 +113,8 @@
       bis-viewerid="#viewer"
       bis-viewerapplicationid="#viewer_application"
       bis-graphelement="#bisweb_graphelement"
-      bis-painttool="#painttool">
-    </bisweb-filetreepanel>
-
-    <bisweb-filetreepipeline
-      id="bis_filetreepipeline"
-      bis-layoutwidgetid="#viewer_layout"
-      bis-filetreepanelid="#bis_filetreepanel"
-    ></bisweb-filetreepipeline>
+      bis-painttool="#painttool"
+    ></bisweb-filetreepanel>
 
     <bisweb-dicomimportelement
       id="dicom_import"
